+++ umami/preprocessing_tools/resampling/count_sampling.py
@@ -314,7 +314,8 @@ class UnderSampling(SimpleSamplingBase):
             logger.info("Plotting resampled distributions...")
             preprocessing_plots(
                 sample=self.config.get_file_name(
-                    option="resampled", use_val=self.use_validation_samples
+                    option="resampled",
+                    use_val=self.use_validation_samples,
                 ),
                 var_dict=get_variable_dict(self.config.config["var_file"]),
                 class_labels=self.config.sampling.class_labels,